fix: renaming new components & tests fixed
