--- components/ImportKey/index.js
@@ -1,3 +1,0 @@
-import ImportKey from './ImportKey';
-
-export default ImportKey;
