+++ components/Ufvk/components/ScannerUfvk.tsx
